docker/requirements.txt
smbus2==0.3.0
grpcio==1.33.1
grpcio-tools==1.33.1